Extract RSF processing code from update_descriptions
This separates out some code I wrote for parsing and generating RSF
into a standalone python package, so I can reuse it for another script.
